project-be-books/pyproject.toml
[tool.poetry]
name = "project-be-books"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
responses = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
